Use a new Manager.get() method to make manager's getters consistent
This is especially useful with `get_or_none` and `get_or_create` methods,
because they can now leverage the argument filling logic used in `all`, and
thus be able to run without specifying all arguments.

That way you can just give them the arguments that are needed to find the
datafile, and then they will load the other values if the file exist, fallback
if not (either by returning `None` or creating the said datafile, as before).